// govco-pwa/src/app/tramites-pwa/models/puntos-de-atencion/puntos-de-atencion-interface.ts
export interface PuntosDeAtencionInterface {
    PuntoAtencionId: string;
    PuntoAtencionNombre: string;
    HorarioAtencion: string;
    PuntoAtencionDireccion: string;
    PuntoAtencionTelefono: string;
    Latitud: string;
    Longitud: string;
    Municipio: string;
    Departamento: string;
}